+++ user/apps/kcc/demo.c
@@ -1,0 +1,6 @@
+#include "../../lib/include/stdio.h"
+
+int main() {
+    printf("Greetings from dynamically compiled program!\n");
+    return 0;
+}